src/backend/requirements.txt
Flask
Flask-SQLAlchemy
flask-jwt-extended
Flask-Cors
python-dotenv
psycopg2-binary
